complete mainsort and creategroups

+++ src/com/group/util/ransort.java
@@ -7,7 +7,7 @@ import java.util.List;
 public class ransort {
 	public List<List<Integer>> randomsort(ArrayList<Integer> studentid_list, int groupnum){
 		Collections.shuffle(studentid_list);
-		int student_per_group = Math.round(studentid_list.size()/groupnum);
+		int student_per_group = Math.round(((float)studentid_list.size())/groupnum);
 		List<List<Integer>> list = new ArrayList<List<Integer>>();
 		int p = 0;
 		for(int i=0; i < groupnum-1; i++){
